+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Maintenance
 * Maintain torch>=1.0.0 only (#39)
 * Manifolds are Modules (#49)
 * Replace deprecated functions in torch>=1.1.0 (#67)
+* Remove PyManOpt from test requirements (#75)
 
 Deprecations
 ------------